% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBioCmirror.R
\name{checkBioCmirror}
\alias{checkBioCmirror}
\title{Check whether a Bioconductor mirror is configured}
\usage{
checkBioCmirror(
  mirror = getOption("BioC_mirror"),
  version = BiocManager::version(),
  repoType = c("BioCsoft", "BioCann", "BioCexp", "BioCworkflows", "BioCbooks"),
  type = getOption("pkgType")
)
}
\arguments{
\item{mirror}{\code{character(1)} The Bioconductor mirror to be tested resolves
from \code{getOption("BioC_mirror")}. If one is not selected,
\code{chooseBioCmirror()} will be run and the user will be able to select a
mirror interactively.}

\item{version}{\code{character(1)} The Bioconductor version to be checked.}

\item{repoType}{\code{character(1)} One of the Bioconductor package types,
including \code{BioCsoft} (default), \code{BioCann}, \code{BioCexp}, \code{BioCworkflows}, and
\code{BioCbooks}.}

\item{type}{\code{character(1)} The package type whether "source", "binary",
"mac.binary", or "win.binary" as given by \code{getOption("pkgType")}. See
\code{install.packages} for details.}
}
\value{
A logical value indicating availability of the mirror indicated in
the name.
}
\description{
The function checks the \code{PACKAGES} file with \code{httr::HEAD}.
}
\examples{

bioc_url <- "https://bioconductor.org"
bioc_mirror <- c(`0-Bioconductor (World-wide) [https]` = bioc_url)
checkBioCmirror(mirror = bioc_mirror)

}
