% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{repositories}
\alias{repositories}
\title{Display Bioconductor and CRAN snapshot repositories}
\usage{
repositories(
  site_repository = character(),
  version = BiocManager::version(),
  snapshot = getOption("BiocArchive.snapshot", "RSPM")
)
}
\arguments{
\item{site_repository}{(Optional) \code{character(1)} representing an
additional repository (e.g., a URL to an organization's
internally maintained repository) in which to look for packages
to install. This repository will be prepended to the default
repositories returned by the function.}

\item{version}{(Optional) \code{character(1)} or \code{package_version}
indicating the \emph{Bioconductor} version (e.g., "3.8") for which
repositories are required.}

\item{snapshot}{\code{character(1)} The snapshot CRAN repository to use for
reproducibility. This defaults to the value of
\code{getOption("BiocArchive.snapshot", "RSPM")}.}
}
\value{
A character vector of Bioconductor and CRAN repositories accounting
for previous releases
}
\description{
\code{repositories()} reports the URLs from which to install \emph{Bioconductor} and
CRAN packages. There repositories are based on the data of the last build
of the archived Bioconductor release.
}
\details{
The CRAN repository reflects the \code{snapshot} arguments, which can be either
\code{RSPM}, \code{MRAN}, or \code{CRAN}. The \code{CRAN} option will default to the established
repository. For installation of archived packages on CRAN, see the
\link{CRANinstall} function.
}
\examples{
if (interactive()) {
    # run within the Bioconductor 3.14 Docker container
    repositories(version = "3.14", snapshot = "RSPM")
}

}
\seealso{
\link{CRANinstall}
}
