% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid}
\alias{valid}
\title{Validate installed packages versions against archive constraints}
\usage{
valid(
  pkgs = utils::installed.packages(lib.loc, priority = priority),
  lib.loc = NULL,
  priority = "NA",
  type = getOption("pkgType"),
  filters = NULL,
  ...,
  checkBuilt = FALSE,
  site_repository = character(),
  version = BiocManager::version(),
  last_built = lastBuilt(version = version),
  snapshot = getOption("BiocArchive.snapshot", "RSPM")
)
}
\arguments{
\item{pkgs}{A character() vector of package names for checking, or
a matrix as returned by \code{\link{installed.packages}()}`.}

\item{lib.loc}{A character() vector of library location(s) of
packages to be validated; see \code{\link{installed.packages}()}.}

\item{priority}{character(1) Check validity of all, "base", or
"recommended" packages; see \code{\link{installed.packages}()}.}

\item{type}{character(1) The type of available package (e.g.,
binary, source) to check validity against; see
\code{\link{available.packages}()}.}

\item{filters}{character(1) Filter available packages to check
validity against; see \code{\link{available.packages}()}.}

\item{...}{Additional arguments, passed to
\code{BiocManager::\link[BiocManager]{install}()} when \code{fix=TRUE}.}

\item{checkBuilt}{\code{logical(1)}. If \code{TRUE} a package built under an
earlier major.minor version of R (e.g., 3.4) is considered to
be old.}

\item{site_repository}{\code{character(1)}. See \code{?install}.}

\item{version}{\code{character(1)} The desired version to reproduce. This is
largely dictated by the current R / Bioconductor version installed and is
indicated by \code{BiocManager::version} by default.}

\item{last_built}{\verb{named character(1)} A character scalar of the date of the
Bioconductor versions last build. The name corresponds to the Bioconductor
version, e.g., \code{c('3.14' = "2022-04-13")}. By default, the \code{lastBuilt()}
function reports the date from the value of the \code{version} argument.}

\item{snapshot}{\code{character(1)} The snapshot CRAN repository to use for
reproducibility. This defaults to the value of
\code{getOption("BiocArchive.snapshot", "RSPM")}.}
}
\value{
\code{biocValid} list object with elements \code{too_new} and \code{out_of_date}
containing \code{data.frame}s with packages and their installed locations that
are too new or out-of-date for the archived version of \emph{Bioconductor}. When
internet access is unavailable, an empty 'biocValid' list is returned. If
all packages ('pkgs') are up to date, then \code{TRUE} is returned.
}
\description{
Check that installed packages are consistent (neither out-of-date nor too
new) with the verison of R and \emph{Bioconductor} in use.
}
\details{
The function mainly works with the CRAN repositories location given
by the \code{repositories()} function. There is no reliable way to verify
\code{CRANinstall} installations other than by running the function with the
\code{dry.run = TRUE} parameter and manually comparing versions.
}
\examples{
if (interactive()) {
  valid()
}
}
