% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocBuild-class.R
\docType{class}
\name{BiocBuild-class}
\alias{BiocBuild-class}
\alias{.BiocBuild}
\alias{BiocBuild}
\alias{show,BiocBuild-method}
\alias{buildDate}
\alias{buildDate,BiocBuild-method}
\alias{version}
\alias{version,BiocBuild-method}
\alias{version<-}
\alias{version<-,BiocBuild-method}
\title{BiocBuild class for indicating Bioconductor build dates}
\usage{
BiocBuild(version = BiocManager::version(), buildDate = Sys.Date())

\S4method{show}{BiocBuild}(object)

buildDate(object)

\S4method{buildDate}{BiocBuild}(object)

version(object)

\S4method{version}{BiocBuild}(object)

version(object) <- value

\S4method{version}{BiocBuild}(object) <- value
}
\arguments{
\item{version}{\code{character(1)} The desired version to reproduce. This is
largely dictated by the current R / Bioconductor version installed and is
indicated by \code{BiocManager::version} by default.}

\item{buildDate}{\code{Date(1)} or \code{character(1)} A scalar vector indicating the
date of the last Bioconductor build.}

\item{object}{A \code{BiocBuild} instance}

\item{value}{\code{character(1)} The Bioconductor \code{version} desired. It is passed
as input to \code{buildDate()}.}
}
\description{
BiocBuild class for indicating Bioconductor build dates
}
\section{Functions}{
\itemize{
\item \code{show(BiocBuild)}: The standard \code{show} method for \code{BiocBuild}

\item \code{buildDate(BiocBuild)}: Extract the \code{buildDate} date from \code{BiocBuild}
objects

\item \code{version(BiocBuild)}: Obtain the Bioconductor version from a
\code{BiocBuild} object

\item \code{version(BiocBuild) <- value}: Change the Bioconductor archive version

}}
\section{Slots}{

\describe{
\item{\code{version}}{A \code{package_version} class indicating the Bioconductor
version}

\item{\code{buildDate}}{The date of the Bioconductor release as a \code{Date} class}
}}

\examples{

BB <- BiocBuild(version = "3.14", buildDate = "2022-04-13")
buildDate(BB)
version(BB)

version(BB) <- "3.13"
BB

}
