% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{CRANinstall}
\alias{CRANinstall}
\title{Install packages from the CRAN archive}
\usage{
CRANinstall(
  pkgs,
  version = BiocManager::version(),
  dry.run = FALSE,
  ...,
  last_built = lastBuilt(version = version)
)
}
\arguments{
\item{pkgs}{\code{character()} A vector of package names whose versions are sought
to be compatible with the archived version of Bioconductor.}

\item{version}{\code{character(1)} The desired version to reproduce. This is
largely dictated by the current R / Bioconductor version installed and is
indicated by \code{BiocManager::version} by default.}

\item{dry.run}{\code{logical(1)} Whether to show only the time machine repository
and forgo the package installation.}

\item{...}{Additional parameters for the \code{BiocManager::install()} function}

\item{last_built}{\verb{named character(1)} A character scalar of the date of the
Bioconductor versions last build. The name corresponds to the Bioconductor
version, e.g., \code{c('3.14' = "2022-04-13")}. By default, the \code{lastBuilt()}
function reports the date from the value of the \code{version} argument.}
}
\value{
Mostly called for its side effect of installing the package from
the CRAN archives that corresponds to the given Bioconductor version.
}
\description{
The function looks through the CRAN archive for each package and finds the
package versions that are compatible with the archived Bioconductor version
using the release date of that Bioconductor version as reported by
\code{lastBuilt}.
}
\examples{

CRANinstall(c("dplyr", "ggplot2"), version = "3.14", dry.run = TRUE)

}
